requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
cryptography==2.6.1
future==0.17.1
pycparser==2.19
python-telegram-bot==11.1.0
six==1.12.0
xkcdpass==1.17.3
